comments and main
