+++ ccam/Makefile
@@ -1,5 +1,7 @@
 main:
-	gcc -Wall -Wextra ucam.c -o ucam
+	gcc -Wall -Os -c main.c
+	gcc -Wall -Os -c ucam.c
+	gcc -Os -o ucam main.o ucam.o
 .PHONY: clean
 clean:
 	rm *.o *.jpg ucam
